Переделал способ запуска тестов в Yandex.Browser, удалил дублирование кода, пофиксил название пакета тестов
