+++ src/test/java/ru/yandex/praktikum/TestConstructor.java
@@ -1,4 +1,4 @@
-package ru.yandex.praktikum.GoogleChromeTests;
+package ru.yandex.praktikum;
 
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
@@ -16,7 +16,10 @@ public class TestConstructor {
     private WebDriver driver;
 
     @Before
-    public void setUp() {
+    public void setUp()
+    {
+        //Расскомментировать для прогона в Yandex.Browser. По умолчанию запускается Google Chrome
+        //System.setProperty("webdriver.chrome.driver", "C:\\yandexdriver\\yandexdriver.exe");
         driver = new ChromeDriver();
     }
