+++ src/test/java/ru/yandex/praktikum/TestProfilePage.java
@@ -1,4 +1,4 @@
-package ru.yandex.praktikum.GoogleChromeTests;
+package ru.yandex.praktikum;
 
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
@@ -14,11 +14,14 @@ import ru.yandex.praktikum.model.*;
 import ru.yandex.praktikum.user_model.User;
 import ru.yandex.praktikum.user_model.UserGenerator;
 
-public class TestLogin {
+public class TestProfilePage {
     private WebDriver driver;
 
     @Before
     public void setUp() {
+
+        //Расскомментировать для прогона в Yandex.Browser. По умолчанию запускается Google Chrome
+        //System.setProperty("webdriver.chrome.driver", "C:\\yandexdriver\\yandexdriver.exe");
         driver = new ChromeDriver();
     }
 
@@ -33,11 +36,12 @@ public class TestLogin {
         driver.quit();
     }
     @Test
-    @DisplayName("Логин через кнопку 'Войти в аккаунт' внизу главной страницы")
-    public void loginViaMainPageLowerLoginButton() {
+    @DisplayName("Открытие профиля авторизованного юзера")
+    public void openProfilePageWithLoggedInUser() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         RegisterPage registerPage = new RegisterPage(driver);
+        ProfilePage profilePage = new ProfilePage(driver);
         User user = UserGenerator.getRandom();
         mainPage.open();
         mainPage.clickUpperAccountButton();
@@ -49,24 +53,22 @@ public class TestLogin {
         registerPage.inputPassword(user.getPassword());
         registerPage.clickRegisterButton();
         Assert.assertTrue("Страница логина после регистрации не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickConstructorButton(); //возврат на главную
-        
-        mainPage.clickLowerLoginButton();
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
         loginPage.inputEmail(user.getEmail());
         loginPage.inputPassword(user.getPassword());
         loginPage.clickLoginButton();
         Assert.assertTrue("Главная страница после логина не открылась", mainPage.isHeaderDisplayed());
         Assert.assertTrue("Кнопка заказа для авторизованного пользователя не отобразилась", mainPage.isMakeOrderDisplayed());
 
-
+        mainPage.clickUpperAccountButton();
+        Assert.assertTrue("Страница профиля не открылась", profilePage.isProfileButtonDisplayed());
     }
     @Test
-    @DisplayName("Логин через кнопку 'Личный кабинет'")
-    public void loginViaMainPageUpperAccountButton() {
+    @DisplayName("Переход по клику на лого на странице профиля")
+    public void checkClickOnLogoAtProfilePage() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         RegisterPage registerPage = new RegisterPage(driver);
+        ProfilePage profilePage = new ProfilePage(driver);
         User user = UserGenerator.getRandom();
         mainPage.open();
         mainPage.clickUpperAccountButton();
@@ -78,24 +80,24 @@ public class TestLogin {
         registerPage.inputPassword(user.getPassword());
         registerPage.clickRegisterButton();
         Assert.assertTrue("Страница логина после регистрации не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickConstructorButton(); //возврат на главную
-        Assert.assertTrue("Главная страница не открылась", mainPage.isHeaderDisplayed());
-        mainPage.clickUpperAccountButton();
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
         loginPage.inputEmail(user.getEmail());
         loginPage.inputPassword(user.getPassword());
         loginPage.clickLoginButton();
         Assert.assertTrue("Главная страница после логина не открылась", mainPage.isHeaderDisplayed());
         Assert.assertTrue("Кнопка заказа для авторизованного пользователя не отобразилась", mainPage.isMakeOrderDisplayed());
 
-
+        mainPage.clickUpperAccountButton();
+        Assert.assertTrue("Страница профиля не открылась", profilePage.isProfileButtonDisplayed());
+        profilePage.clickLogo();
+        Assert.assertTrue("Главная страница не открылась", mainPage.isHeaderDisplayed());
     }
     @Test
-    @DisplayName("Логин через кнопку 'Войти' на странице регистрации")
-    public void loginViaRegisterPageButton() {
+    @DisplayName("Переход по клику на лого на странице профиля")
+    public void checkClickOnConstructorButtonAtProfilePage() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         RegisterPage registerPage = new RegisterPage(driver);
+        ProfilePage profilePage = new ProfilePage(driver);
         User user = UserGenerator.getRandom();
         mainPage.open();
         mainPage.clickUpperAccountButton();
@@ -107,29 +109,24 @@ public class TestLogin {
         registerPage.inputPassword(user.getPassword());
         registerPage.clickRegisterButton();
         Assert.assertTrue("Страница логина после регистрации не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickConstructorButton(); //возврат на главную
-        Assert.assertTrue("Главная страница не открылась", mainPage.isHeaderDisplayed());
-
-        mainPage.clickUpperAccountButton();
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickRegisterButton();
-        Assert.assertTrue("Страница регистрации не открылась", registerPage.isHeaderDisplayed());
-        registerPage.clickLoginButton();  //Нажимаем кнопку "Войти" на странице регистрации
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
         loginPage.inputEmail(user.getEmail());
         loginPage.inputPassword(user.getPassword());
         loginPage.clickLoginButton();
         Assert.assertTrue("Главная страница после логина не открылась", mainPage.isHeaderDisplayed());
         Assert.assertTrue("Кнопка заказа для авторизованного пользователя не отобразилась", mainPage.isMakeOrderDisplayed());
 
+        mainPage.clickUpperAccountButton();
+        Assert.assertTrue("Страница профиля не открылась", profilePage.isProfileButtonDisplayed());
+        profilePage.clickConstructorButton();
+        Assert.assertTrue("Главная страница не открылась", mainPage.isHeaderDisplayed());
     }
     @Test
-    @DisplayName("Логин через кнопку 'Забыли пароль?'")
-    public void loginViaForgotPasswordPageButton() {
+    @DisplayName("Разлогин по клику на кнопку 'Выход' на странице профиля")
+    public void checkLogoutButtonAtProfilePage() {
         MainPage mainPage = new MainPage(driver);
         LoginPage loginPage = new LoginPage(driver);
         RegisterPage registerPage = new RegisterPage(driver);
-        ForgotPasswordPage forgotPasswordPage = new ForgotPasswordPage(driver);
+        ProfilePage profilePage = new ProfilePage(driver);
         User user = UserGenerator.getRandom();
         mainPage.open();
         mainPage.clickUpperAccountButton();
@@ -141,19 +138,17 @@ public class TestLogin {
         registerPage.inputPassword(user.getPassword());
         registerPage.clickRegisterButton();
         Assert.assertTrue("Страница логина после регистрации не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickConstructorButton(); //возврат на главную
-        Assert.assertTrue("Главная страница не открылась", mainPage.isHeaderDisplayed());
-
-        mainPage.clickUpperAccountButton();
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
-        loginPage.clickForgotPasswordButton();
-        Assert.assertTrue("Страница восстановления пароля не открылась", forgotPasswordPage.isHeaderDisplayed());
-        forgotPasswordPage.clickLoginButton();
-        Assert.assertTrue("Страница логина не открылась", loginPage.isHeaderDisplayed());
         loginPage.inputEmail(user.getEmail());
         loginPage.inputPassword(user.getPassword());
         loginPage.clickLoginButton();
         Assert.assertTrue("Главная страница после логина не открылась", mainPage.isHeaderDisplayed());
         Assert.assertTrue("Кнопка заказа для авторизованного пользователя не отобразилась", mainPage.isMakeOrderDisplayed());
+
+        mainPage.clickUpperAccountButton();
+        Assert.assertTrue("Страница профиля не открылась", profilePage.isProfileButtonDisplayed());
+        profilePage.clickLogoutButton();
+        Assert.assertTrue("Страница логина после разлогина не открылась", loginPage.isHeaderDisplayed());
     }
+
+
 }
